third_party/WebKit/LayoutTests/fast/css/namespaces/invalid-url-in-namespace.html

<!DOCTYPE html>

<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>

<style>
  @namespace html url(http://www.w3.org/1999/xhtml);
  @namespace html url("bad-string-token
  );
  div {
      background-color: red;
      width: 100px;
      height: 100px;
  }
  html|div {
      background-color: green;
  }
</style>

<div id=div></div>

<script>
test(() => {
    assert_equals(getComputedStyle(div).backgroundColor, "rgb(0, 128, 0)");
});
</script>
